move all public trampolines into their own cpp file
This should prevent code duplication due to agressive inlining.
