+++ filament/src/Fence.cpp
@@ -1,0 +1,31 @@
+/*
+ * Copyright (C) 2016 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "details/Fence.h"
+
+namespace filament {
+
+using namespace backend;
+
+FenceStatus Fence::waitAndDestroy(Fence* fence, Mode mode) {
+    return FFence::waitAndDestroy(upcast(fence), mode);
+}
+
+FenceStatus Fence::wait(Mode mode, uint64_t timeout) {
+    return upcast(this)->wait(mode, timeout);
+}
+
+} // namespace filament
