+++ filament/src/details/Renderer.cpp
@@ -1249,66 +1249,4 @@ void FRenderer::initializeClearFlags() {
             | TargetBufferFlags::DEPTH_AND_STENCIL;
 }
 
-
-// ------------------------------------------------------------------------------------------------
-// Trampoline calling into private implementation
-// ------------------------------------------------------------------------------------------------
-
-Engine* Renderer::getEngine() noexcept {
-    return &upcast(this)->getEngine();
-}
-
-void Renderer::render(View const* view) {
-    upcast(this)->render(upcast(view));
-}
-
-bool Renderer::beginFrame(SwapChain* swapChain, uint64_t vsyncSteadyClockTimeNano) {
-    return upcast(this)->beginFrame(upcast(swapChain), vsyncSteadyClockTimeNano);
-}
-
-void Renderer::copyFrame(SwapChain* dstSwapChain, filament::Viewport const& dstViewport,
-        filament::Viewport const& srcViewport, CopyFrameFlag flags) {
-    upcast(this)->copyFrame(upcast(dstSwapChain), dstViewport, srcViewport, flags);
-}
-
-void Renderer::readPixels(uint32_t xoffset, uint32_t yoffset, uint32_t width, uint32_t height,
-        PixelBufferDescriptor&& buffer) {
-    upcast(this)->readPixels(xoffset, yoffset, width, height, std::move(buffer));
-}
-
-void Renderer::readPixels(RenderTarget* renderTarget,
-        uint32_t xoffset, uint32_t yoffset, uint32_t width, uint32_t height,
-        PixelBufferDescriptor&& buffer) {
-    upcast(this)->readPixels(upcast(renderTarget),
-            xoffset, yoffset, width, height, std::move(buffer));
-}
-
-void Renderer::endFrame() {
-    upcast(this)->endFrame();
-}
-
-double Renderer::getUserTime() const {
-    return upcast(this)->getUserTime().count();
-}
-
-void Renderer::resetUserTime() {
-    upcast(this)->resetUserTime();
-}
-
-void Renderer::setDisplayInfo(const DisplayInfo& info) noexcept {
-    upcast(this)->setDisplayInfo(info);
-}
-
-void Renderer::setFrameRateOptions(FrameRateOptions const& options) noexcept {
-    upcast(this)->setFrameRateOptions(options);
-}
-
-void Renderer::setClearOptions(const ClearOptions& options) {
-    upcast(this)->setClearOptions(options);
-}
-
-void Renderer::renderStandaloneView(View const* view) {
-    upcast(this)->renderStandaloneView(upcast(view));
-}
-
 } // namespace filament
