+++ filament/src/details/Scene.cpp
@@ -454,56 +454,4 @@ void FScene::forEach(Invocable<void(utils::Entity)>&& functor) const noexcept {
     std::for_each(mEntities.begin(), mEntities.end(), std::move(functor));
 }
 
-// ------------------------------------------------------------------------------------------------
-// Trampoline calling into private implementation
-// ------------------------------------------------------------------------------------------------
-
-void Scene::setSkybox(Skybox* skybox) noexcept {
-    upcast(this)->setSkybox(upcast(skybox));
-}
-
-Skybox* Scene::getSkybox() noexcept {
-    return upcast(this)->getSkybox();
-}
-
-Skybox const* Scene::getSkybox() const noexcept {
-    return upcast(this)->getSkybox();
-}
-
-void Scene::setIndirectLight(IndirectLight const* ibl) noexcept {
-    upcast(this)->setIndirectLight(upcast(ibl));
-}
-
-void Scene::addEntity(Entity entity) {
-    upcast(this)->addEntity(entity);
-}
-
-void Scene::addEntities(const Entity* entities, size_t count) {
-    upcast(this)->addEntities(entities, count);
-}
-
-void Scene::remove(Entity entity) {
-    upcast(this)->remove(entity);
-}
-
-void Scene::removeEntities(const Entity* entities, size_t count) {
-    upcast(this)->removeEntities(entities, count);
-}
-
-size_t Scene::getRenderableCount() const noexcept {
-    return upcast(this)->getRenderableCount();
-}
-
-size_t Scene::getLightCount() const noexcept {
-    return upcast(this)->getLightCount();
-}
-
-bool Scene::hasEntity(Entity entity) const noexcept {
-    return upcast(this)->hasEntity(entity);
-}
-
-void Scene::forEach(Invocable<void(utils::Entity)>&& functor) const noexcept {
-    upcast(this)->forEach(std::move(functor));
-}
-
 } // namespace filament
