+++ filament/src/details/Skybox.cpp
@@ -151,28 +151,4 @@ void FSkybox::commit(backend::DriverApi& driver) noexcept {
     mSkyboxMaterialInstance->commit(driver);
 }
 
-// ------------------------------------------------------------------------------------------------
-// Trampoline calling into private implementation
-// ------------------------------------------------------------------------------------------------
-
-void Skybox::setLayerMask(uint8_t select, uint8_t values) noexcept {
-    upcast(this)->setLayerMask(select, values);
-}
-
-uint8_t Skybox::getLayerMask() const noexcept {
-    return upcast(this)->getLayerMask();
-}
-
-float Skybox::getIntensity() const noexcept {
-    return upcast(this)->getIntensity();
-}
-
-void Skybox::setColor(math::float4 color) noexcept {
-    upcast(this)->setColor(color);
-}
-
-Texture const* Skybox::getTexture() const noexcept {
-    return upcast(this)->getTexture();
-}
-
 } // namespace filament
