+++ filament/src/details/Stream.cpp
@@ -151,33 +151,4 @@ int64_t FStream::getTimestamp() const noexcept {
     return driver.getStreamTimestamp(mStreamHandle);
 }
 
-// ------------------------------------------------------------------------------------------------
-// Trampoline calling into private implementation
-// ------------------------------------------------------------------------------------------------
-
-StreamType Stream::getStreamType() const noexcept {
-    return upcast(this)->getStreamType();
-}
-
-void Stream::setAcquiredImage(void* image, Callback callback, void* userdata) noexcept {
-    upcast(this)->setAcquiredImage(image, callback, userdata);
-}
-
-void Stream::setAcquiredImage(void* image, backend::CallbackHandler* handler, Callback callback, void* userdata) noexcept {
-    upcast(this)->setAcquiredImage(image, handler, callback, userdata);
-}
-
-void Stream::setDimensions(uint32_t width, uint32_t height) noexcept {
-    upcast(this)->setDimensions(width, height);
-}
-
-void Stream::readPixels(uint32_t xoffset, uint32_t yoffset, uint32_t width, uint32_t height,
-        backend::PixelBufferDescriptor&& buffer) noexcept {
-    upcast(this)->readPixels(xoffset, yoffset, width, height, std::move(buffer));
-}
-
-int64_t Stream::getTimestamp() const noexcept {
-    return upcast(this)->getTimestamp();
-}
-
 } // namespace filament
